staticfiles/staticfiles/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
az-iranian-bank-gateways==2.0.8
certifi==2024.7.4
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
dj-database-url==2.2.0
Django==5.0.7
django-allauth==64.1.0
django-anymail==11.1
django-crispy-forms==2.3
gunicorn==23.0.0
idna==3.8
isodate==0.6.1
lxml==5.3.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycryptodome==3.20.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
zeep==4.2.1
